+++ src/htmx_nav/__init__.py
@@ -1,6 +1,6 @@
 from .partials import PartialSpec
 from .shell import ShellRenderer, make_shell_renderer
-from .shortcuts import render_nav, render_with_swaps
+from .shortcuts import render_nav
 from .swaps import Swap, Swaps
 from .targeting import Target, has_messages, htmx_target_is, not_targeting, targeting
 from .views import make_shell_view_mixin
@@ -16,7 +16,6 @@ __all__ = [
     "not_targeting",
     "has_messages",
     "render_nav",
-    "render_with_swaps",
     "make_shell_renderer",
     "make_shell_view_mixin",
 ]